+++ packages/vite-plugin-cloudflare/playground/prisma/src/index.ts
@@ -1,12 +1,12 @@
 import { PrismaD1 } from "@prisma/adapter-d1";
-import { PrismaClient } from "@prisma/client-generated";
+import { PrismaClient } from "./generated/prisma/client";
 
 interface Env {
 	DB: D1Database;
 }
 
 export default {
-	async fetch(request, env) {
+	async fetch(_, env) {
 		const adapter = new PrismaD1(env.DB);
 		const prisma = new PrismaClient({ adapter });
 		const users = await prisma.user.findMany();