+++ packages/vite-plugin-cloudflare/src/shared.ts
@@ -10,6 +10,8 @@ export const additionalModuleGlobalRE = new RegExp(
 	"g"
 );
 
+// headers
 export const WORKER_ENTRY_PATH_HEADER = "__VITE_WORKER_ENTRY_PATH__";
+export const IS_ENTRY_WORKER_HEADER = "__VITE_IS_ENTRY_WORKER__";
 
 export const VIRTUAL_WORKER_ENTRY = `${virtualPrefix}worker-entry`;