+++ packages/wrangler/src/__tests__/config/configuration.test.ts
@@ -129,6 +129,8 @@ describe("normalizeAndValidateConfig()", () => {
 			usage_model: undefined,
 			vars: {},
 			define: {},
+			definedEnvironments: [],
+			targetEnvironment: undefined,
 			wasm_modules: undefined,
 			data_blobs: undefined,
 			workers_dev: undefined,